Added detection of vulnerable ImageMagick
Added a new template named "imagetragick" that detects if the file
upload is vulnerable to CVE-2016-3714, which is a Remote Code Execution
vulnerability in the ImageMagick software. By uploading a crafted file
it is possible to run arbitrary commands on the server if vulnerable.
The command performed by the template on vulnerable servers creates a
file with the text "ImageTragick Detected!" in the same directory as the
upload form page.

Since no malicious extension is needed for the exploitation of this
vulnerability, support has been added for templates without "nasty"
extensions. This vulnerability can be exploited by uploading a variety
of image filetypes (png, jpg etc.), so all legitimate filetypes
supported by the file upload in the server will be tried.

Tested against a docker machine running a vulnerable instance of
ImageMagick. For testing, created a variant of a docker machine created
by the chinese VulApps: http://vulapps.evalbug.com/i_imagemagick_1/
The docker variant used for testing can be found by running:
- docker pull madhatter37/vulnerable_apps:imagemagick_1.0.2
- docker run -d -p 80:80 madhatter37/vulnerable_apps:imagemagick_1.0.2
- ./fuxploider.py --uploads-path uploads -u http://127.0.0.1/file_upload/poc.php --not-regex "your file was not uploaded" --true-regex "has been uploaded to" --data "submit=Submit" --template=imagetragick