Removed requirement for Upload Folder for templates with "codeExecURL"
Templates which provide a custom "codeExecURL" parameter do not
require the "--uploads-path" parameter, since they provide a custom
URL for checking execution (can depend on form URL and/or filename).
Now they should function properly also if the parameter isn't set.

Updated the warnings to include an exception for templates with
"codeExecURL" parameter.