New template: upload .htaccess
This template attempts to exploit a misconfiguration in an apache server.
If the apache server allows to override the default ".htaccess" file by
enabling the "AllowOverride" parameter for the upload directory, then it
is possible to exploit this configuration for PHP code execution.

In order to implement the template, a new parameter was added to the
"templates" structure named "staticFilename". When enabled, all payloads
will be uploaded with the same filename (including suffix). Only mime
types will be altered between requests. This is enabled on the
"htaccess" template, since the file with be ignored if uploaded with any
other suffix.

This template was tested against an Apache 2.4 server (notice the syntax
differs for Apache 2.2 servers). The test was performed by the following
commands:
sudo docker pull madhatter37/vulapps:file_upload
sudo docker run -d -p 80:80 --name=file_upload  madhatter37/vulapps:file_upload
./fuxploider.py -t htaccess -u http://127.0.0.1/ --true-regex "has been uploaded" --uploads-path uploads